fix(PollManager): change index type from int to size_t in removeFd method
- Updated the removeFd method to use size_t for the index variable, improving type safety and consistency with the vector size type.

+++ src/server/PollManager.cpp
@@ -34,7 +34,7 @@ void PollManager::addFd(int fd, short events) {
 };
 
 void PollManager::removeFd(int fd) {
-	int idx = findIndex(fd);
+	size_t idx = findIndex(fd);
 	if (idx < _fds.size()) {
 		_fds.erase(_fds.begin() + idx);
 	}
